+++ user-service/prototype/src/main/java/jp/co/itmeister/userservice/userservice/dto/PostDto.java
@@ -18,8 +18,8 @@ public class PostDto {
     private String description;
     private String brand;
     private String cameraName;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
     private ZonedDateTime snapTime;
     private Integer angle;
     private String iso;
@@ -33,7 +33,7 @@ public class PostDto {
     // コンストラクタ
     public PostDto ( Integer userId, String url, String title, Short prefecture, 
                          String cityName, String description, String brand, String cameraName, 
-                         BigDecimal latitude, BigDecimal longitude, ZonedDateTime snapTime, 
+                         double latitude, double longitude, ZonedDateTime snapTime, 
                          Integer angle, String iso, BigDecimal fValue, String shutterSpeed) {
         this.userId = userId;
         this.url = url;
@@ -116,19 +116,19 @@ public class PostDto {
         this.cameraName = cameraName;
     }
 
-    public BigDecimal getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(BigDecimal latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public BigDecimal getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(BigDecimal longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
