+++ user-service/prototype/src/main/java/jp/co/itmeister/userservice/userservice/repository/PostRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import jp.co.itmeister.userservice.userservice.entity.PostEntity;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -23,8 +22,8 @@ public interface PostRepository extends JpaRepository<PostEntity, Long> {
                    "ORDER BY function('ST_Distance', p.latlng, " +
                    "function('ST_SetSRID', function('ST_MakePoint', :longitude, :latitude), 4326))" +
                    "LIMIT :limit")
-    List<PostEntity> findNearbyPosts(@Param("latitude") BigDecimal latitude,
-                                     @Param("longitude") BigDecimal longitude,
+    List<PostEntity> findNearbyPosts(@Param("latitude") double latitude,
+                                     @Param("longitude") double longitude,
                                      @Param("radius") double searchRadius,
                                      @Param("limit") Integer limit);
 }
